requirements.txt
h11>=0.16.0
pip>=24.0
distro>=1.9.0
Jinja2>=3.1.6
typing_extensions>=4.15.0
anyio>=4.12.1
sniffio>=1.3.1
pydantic>=2.12.5
httpx>=0.28.1
pydantic_core>=2.41.5
PyYAML>=6.0.3
httpcore>=1.0.9
idna>=3.11
zstandard>=0.25.0
MarkupSafe>=3.0.3
orjson>=3.11.7
certifi>=2026.1.4
urllib3>=2.6.3
tenacity>=9.1.4
langchain>=1.2.10
langgraph>=1.0.8
xxhash>=3.6.0
langsmith>=0.7.1
ormsgpack>=1.12.2
requests>=2.32.5
uuid_utils>=0.14.0
packaging>=26.0
jsonpointer>=3.0.0
jsonpatch>=1.33
python-dotenv>=1.2.1